separate devices and telemetry views
