+++ backend/apps/devices/views/__init__.py
@@ -1,0 +1,4 @@
+from .device_views import DeviceView, DeviceDetailView
+from .telemetry_views import ingest_telemetry
+
+__all__ = ['DeviceView', 'DeviceDetailView', 'ingest_telemetry']
